refactor(web): share the formatters and the treemap layout
`compactNumber`, `fileName` and `dirName` sat in `lib/hotspots` because the
hotspot screen was the only screen. The dependency graph needs all three, and
a feature folder importing another feature's internals is the wrong shape, so
they move up into `lib/format`.

`squarify` moves the same way, into `lib/geometry` — it is geometry, not a
hotspot, and naming it for what it is says so.

No behaviour changes; the hotspot views import from the new homes.
