+++ apps/web/src/lib/format/number.ts
@@ -1,7 +1,8 @@
 /**
- * Hotspot scores run from single digits into the millions on a large repo. The
- * table has to stay scannable at both ends, so anything past a thousand is
- * abbreviated and only the leading digits are kept.
+ * Analysis numbers run from single digits into the millions on a large repo —
+ * hotspot scores, lines of code, edge counts. A table or a stat panel has to
+ * stay scannable at both ends, so anything past a thousand is abbreviated and
+ * only the leading digits are kept.
  */
 export function compactNumber(value: number): string {
   if (!Number.isFinite(value)) return '—';
