+++ apps/web/src/lib/geometry/index.ts
@@ -1,0 +1,1 @@
+export { squarify, type TreemapTile } from './squarify';
