+++ apps/web/src/lib/geometry/squarify.ts
