+++ apps/web/src/lib/hotspots/index.ts
@@ -1,4 +1,3 @@
-export { compactNumber } from './format';
 export {
   heatBands,
   heatColor,
@@ -10,6 +9,4 @@ export {
   type HeatLevel,
   type HeatScale,
 } from './heat';
-export { dirName, fileName } from './path';
 export { hotspotRows, HOTSPOT_SERIES_ID, type HotspotRow } from './rows';
-export { squarify, type TreemapTile } from './layout';
